Seriously unfucked mod structure. Apparently individual mods have to be in "Seperate Folders" or something SMH
