+++ 121st_MSV/cfgPatches.hpp
@@ -4,6 +4,7 @@ class CfgPatches
 {
 	class 121st_MSV
 	{
+		addonRootClass = "Valhalla";
 		units[]=
 		{
 			"121MSV_Soldier_base_f", "121MSV_Platoon_Leader", "121MSV_Platoon_Medic", "121MSV_Section_Commander", "121MSV_Assistant_Grenadier", "121MSV_Machine_Gunner", "121MSV_Assistant_Machine_Gunner", "121MSV_Rifleman", "121MSV_Senior_Rifleman", "121MSV_Grenadier", "121MSV_VehicleCrew", "121MSV_BMP2D"
@@ -13,14 +14,5 @@ class CfgPatches
 		};
 		requiredVersion=0.1;
 		requiredAddons[]={"A3_Characters_F"};
-	};
-	class 75th_RR
-	{
-		units[]=
-		{
-			"75th_RR_Soldier_base_f", "75th_RR_Platoon_Leader", "75th_RR_Medic", "75th_RR_Platoon_Sgt", "75th_RR_JTAC", "75th_RR_Squad_Leader", "75th_RR_FTL", "75th_RR_Rifleman", "75th_RR_AR", "75th_RR_Grenadier", "75th_RR_MG", "75th_RR_AMG", "75th_RR_AT", "75th_RR_AAT"
-		};
-		requiredVersion=0.1;
-		requiredAddons[]={A3_Characters_F"};
 	};
 };
