+++ 121st_MSV/config.cpp
@@ -1,0 +1,8 @@
+/// Utilizing includes ///
+//Github repo VALHALLA
+	#include "defines.hpp"
+	#include "cfgPatches.hpp"
+	#include "cfgFactionClasses.hpp"
+	#include "cfgWeapons.hpp"
+	#include "cfgVehicles.hpp"
+	#include "cfgUnits.hpp"
