+++ 75th_RR/cfgFactionClasses.hpp
@@ -1,0 +1,10 @@
+class CfgFactionClasses
+{
+	class 75th_RR
+	{
+		displayName="75th Ranger Regiment";
+		priority=4;
+		side=1;
+		icon= "";
+	};
+};
