+++ 75th_RR/cfgPatches.hpp
@@ -1,0 +1,15 @@
+/* TASK FORCE 121 VALHALLA MOD CLASS DEFINITIONS*/
+
+class CfgPatches
+{
+	class 75th_RR
+	{
+		addonRootClass = "Valhalla";
+		units[]=
+		{
+			"75th_RR_Soldier_base_f", "75th_RR_Platoon_Leader", "75th_RR_Medic", "75th_RR_Platoon_Sgt", "75th_RR_JTAC", "75th_RR_Squad_Leader", "75th_RR_FTL", "75th_RR_Rifleman", "75th_RR_AR", "75th_RR_Grenadier", "75th_RR_MG", "75th_RR_AMG", "75th_RR_AT", "75th_RR_AAT"
+		};
+		requiredVersion=0.1;
+		requiredAddons[]={A3_Characters_F"};
+	};
+};
