+++ 75th_RR/cfgWeapons.hpp
@@ -1,0 +1,1 @@
+//no weapons yet...
