+++ 75th_RR/config.cpp
@@ -1,0 +1,9 @@
+/// Utilizing includes ///
+//Github repo VALHALLA
+
+	#include "defines.hpp"
+	#include "cfgPatches.hpp"
+	#include "cfgFactionClasses.hpp"
+	#include "cfgWeapons.hpp"
+	#include "cfgVehicles.hpp"
+	#include "cfgUnits.hpp"
