+++ 75th_RR/defines.hpp
@@ -1,0 +1,18 @@
+/// Define Magazine Handler Macros///
+
+#define mag_2(a) a, a
+#define mag_3(a) a, a, a
+#define mag_4(a) a, a, a, a
+#define mag_5(a) a, a, a, a, a
+#define mag_6(a) a, a, a, a, a, a
+#define mag_7(a) a, a, a, a, a, a, a
+#define mag_8(a) a, a, a, a, a, a, a, a
+#define mag_9(a) a, a, a, a, a, a, a, a, a
+#define mag_10(a) a, a, a, a, a, a, a, a, a, a
+#define mag_11(a) a, a, a, a, a, a, a, a, a, a, a
+#define mag_12(a) a, a, a, a, a, a, a, a, a, a, a, a
+#define mag_xx(a,b) class _xx_##a {magazine = a; count = b;}
+#define weap_xx(a,b) class _xx_##a {weapon = a; count = b;}
+#define item_xx(a,b) class _xx_##a {name = a; count = b;}
+#define IFAK mag_7(ACE_quikclot), mag_2(ACE_tourniquet),mag_5(ACE_packingBandage),mag_3(ACE_elasticBandage),mag_2(ACE_EarPlugs)
+#define CommonStuff ItemMap, ItemCompass, ItemWatch, ItemRadio, rhs_1PN138, ACE_MapTools
