--- cfgVehicles.hpp
@@ -1,7 +1,0 @@
-#include defines.hpp
-
-class cfgVehicles		// Character classes are defined under cfgVehicles.
-{
-	#include cfgVehicles_MSV
-	#include cfgVehicles_75th
-};
