+++ config.cpp
@@ -1,8 +1,20 @@
-/// Utilizing includes ///
-//Github repo VALHALLA
+class CfgPatches
+{
+	class Valhalla
+	{
+		// Meta information for editor
+		name = "Valhalla";
+		author = "Task Force 121 Valhalla Team";
+		url = "http://www.taskforce121.us";
 
-	#include "cfgPatches.hpp"
-	#include "cfgFactionClasses.hpp"
-	#include "cfgWeapons.hpp"
-	#include "cfgVehicles.hpp"
-	#include "cfgUnits.hpp"
+		// Minimum compatible version. When the game's version is lower, pop-up warning will appear when launching the game. Note: was disabled on purpose some time late into Arma 2: OA.
+		requiredVersion = 1.60; 
+		// Required addons, used for setting load order.
+		// When any of the addons is missing, pop-up warning will appear when launching the game.
+		requiredAddons[] = { "A3_Functions_F" };
+		// List of objects (CfgVehicles classes) contained in the addon. Important also for Zeus content (units and groups) unlocking.
+		units[] = {};
+		// List of weapons (CfgWeapons classes) contained in the addon.
+		weapons[] = {};
+	};
+};
